web/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.5.0
asyncpg==0.29.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
